pyproject.toml
[tool.poetry]
name = "weartheapi"
version = "0.1.0"
description = ""
authors = ["fertlopus <marat.movlamov@icloud.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.36"
alembic = "^1.13.3"
asyncpg = "^0.30.0"
psycopg2-binary = "^2.9.10"
redis = "^5.2.0"
python-dotenv = "^1.0.1"
azure-identity = "^1.19.0"
azure-keyvault-secrets = "^4.9.0"
pydantic-settings = "^2.6.1"
aioredis = "^2.0.1"


[tool.poetry.group.dev.dependencies]
openai = "^1.51.0"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
llama-index = "^0.11.16"
langchain = "^0.3.2"
pytest = "^8.3.3"
pytest-cov = "^6.0.0"
black = "^24.10.0"
flake8 = "^7.1.1"
pre-commit = "^4.0.1"
scikit-learn = "^1.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 140
target-version = ['py310', 'py311', 'py312']
skip-string-normalization = true
skip-magic-trailing-comma = true
